Move email-related components into the lib/service/email directory.
