+++ app/lib/frontend/handlers/report.dart
@@ -14,12 +14,12 @@ import 'package:shelf/shelf.dart' as shelf;
 import '../../../service/rate_limit/rate_limit.dart';
 import '../../account/backend.dart';
 import '../../admin/models.dart';
-import '../../frontend/email_sender.dart';
 import '../../frontend/handlers/cache_control.dart';
 import '../../package/backend.dart';
 import '../../publisher/backend.dart';
+import '../../service/email/email_sender.dart';
+import '../../service/email/email_templates.dart';
 import '../../shared/datastore.dart';
-import '../../shared/email.dart';
 import '../../shared/exceptions.dart';
 import '../../shared/handlers.dart';
 import '../request_context.dart';
