+++ app/lib/package/backend.dart
@@ -32,11 +32,11 @@ import '../account/models.dart' show User;
 import '../audit/models.dart';
 import '../publisher/backend.dart';
 import '../service/email/backend.dart';
+import '../service/email/email_templates.dart';
 import '../service/email/models.dart';
 import '../service/secret/backend.dart';
 import '../shared/configuration.dart';
 import '../shared/datastore.dart';
-import '../shared/email.dart';
 import '../shared/exceptions.dart';
 import '../shared/redis_cache.dart' show cache;
 import '../shared/storage.dart';
