+++ app/test/shared/email_test.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:pub_dev/shared/email.dart';
+import 'package:pub_dev/service/email/email_templates.dart';
 import 'package:test/test.dart';
 
 void main() {
